Check for hitting through walls
